Refactors paths
