+++ rascal2/widgets/plot.py
@@ -8,7 +8,8 @@ import ratapi
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg, NavigationToolbar2QT
 from PyQt6 import QtCore, QtGui, QtWidgets
 
-from rascal2.config import SETTINGS, path_for
+from rascal2.config import SETTINGS
+from rascal2.paths import path_for
 from rascal2.widgets.inputs import MultiSelectComboBox, ProgressButton
